+++ openai_assistant/test/test_files_handler.py
@@ -3,23 +3,34 @@ import unittest
 from typing import List, Iterable
 
 from dotenv import load_dotenv
+from openai._types import FileTypes
 from openai.types import FileObject
 
-# from utils import files_handler
-
 load_dotenv('../.env', override=True)
+from utils.files_handler import files_handler  # noqa: E402
 TEST_FILE_ID = os.getenv('TEST_FILE_ID')
 
 
-@unittest.skip("Fails to import client in the files_handler.py file")
 class TestFilesHandler(unittest.IsolatedAsyncioTestCase):
 
+    async def test_init(self):
+        await files_handler.init()
+
     async def test_files_list(self):
-        files: List[FileObject] = await files_handler.files_list()
+        files: List[FileObject] = await files_handler._files_list()
         self.assertTrue(isinstance(files, Iterable), "ojb should be an iterable")
         self.assertTrue(all(isinstance(item, FileObject) for item in files),
                         "all items in files should be of type FileObject")
 
     async def test_file_retrieve(self):
-        result = await files_handler.files_retrieve(TEST_FILE_ID)
+        result = await files_handler._files_retrieve(TEST_FILE_ID)
         self.assertTrue(isinstance(result, FileObject), "ojb should be an FileObject")
+
+    @unittest.skip('Side effects')
+    async def test_file_create(self):
+
+        file_name = 'test/dummy_file.txt'
+        file_content = bytes(100)
+        file: FileTypes = (file_name, file_content)
+        result = await files_handler.create(file)
+        self.assertTrue(result.filename, file_name)